+++ lib/sugar_refinery/file_force_delete.rb
@@ -1,0 +1,12 @@
+require_relative 'version'
+
+module SugarRefinery
+  module FileForceDelete
+    refine File.singleton_class do
+      def delete!(filename)
+        return nil if !File.exist?(filename)
+        File.delete filename
+      end
+    end
+  end
+end